+++ StaticPopup.lua
@@ -7,7 +7,7 @@ local L = LibStub("AceLocale-3.0"):GetLocale("FarmingBar", true)
 -- TODO: Localize text
 
 StaticPopupDialogs["FARMINGBAR_CONFIRM_DELETE_MULTIPLE_OBJECTIVES"] = {
-    text = "You are about to delete %d objectives. |cffff0000This will remove these objectives from all bars globally.|r Do you want to continue?",
+    text = "You are about to delete %d objectives. |cffff0000This will remove these objectives from all bars globally.|r These objectives are active on %d button(s). Do you want to continue?",
     button1 = YES,
     button2 = NO,
     OnAccept = function()
@@ -22,7 +22,7 @@ StaticPopupDialogs["FARMINGBAR_CONFIRM_DELETE_MULTIPLE_OBJECTIVES"] = {
 ------------------------------------------------------------
 
 StaticPopupDialogs["FARMINGBAR_CONFIRM_DELETE_OBJECTIVE"] = {
-    text = [[You are about to delete the objective "%s". |cffff0000This will remove this objective from all bars globally.|r Do you want to continue?]],
+    text = [[You are about to delete the objective "%s". |cffff0000This will remove this objective from all bars globally.|r This objective is active on %d button(s). Do you want to continue?]],
     button1 = YES,
     button2 = NO,
     OnAccept = function(_, objectiveTitle)